aws_sns_middleware/pyproject.toml
[tool.poetry]
name = "aws-sns-middleware"
version = "0.1.0"
description = ""
authors = ["Christopher Kosik <37384678+skorm11x@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
boto3 = "^1.34.152"
python-dotenv = "^1.0.1"
pylint = "^3.2.6"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
